Implemented sharing data between experiments specifically inheriting data from previous experiment. Works for normal tree experiments and the experiment iterator.
nv_loc_experiment.py shows how to use.

Requirments:
 - new experiment has 'inherit_data' in settings with value of True
 - new experiment sets the keys of desired inherited data. This is done in the __init__ with something as simple as self.data[key] = None

In main_window.py and experiment_iterator.py, there is a check for both inheit data is true and the key matching. If both conditions are met it updates the key of the new experiment to have the same value as the experiment ran before it. If there was no experiment beforehand nothing happens.