Iterator working with 1,2, and 3 layers tested!! Should be up to N layers with the recursive logic!
Still need to incorporate the sub iterator settings although they might not be important.

Anyway the next step is to make sure the data shapes are the same and if they are not go through all the data to reshape ://. But at least the cases of uniform data is working!