Changed up plotting in experiment_iterator.py.
With new data structure the old method had an issue where the iterators data dictionary is not set up the same as the data dictionary for normal experiments.

When the iterator is running, it plots the same as the current running experiment. After running, for a loop iterator it tries to plot the iterator's average data. If that fails it defaults to plotting the last ran experiment's data.

For a sweep iterator when running it plots the same as the current running experiment. After running it plots the last ran experiment's data. Since the data structure is complex this seems like the best approcach. Proper data analysis should be done with saved data.

Added a _plot method to iterator class that basically mirros plot as _plot is typically called last and if there is a NotImplemented Error, the plot is blank.